requirements.txt
fastapi==0.81.0,<0.82.0
requests==2.28.1,<2.28.2
tensorflow==2.9.1,<2.10
uvicorn==0.18.3,<0.19
spliceai==1.3.1,<1.4
PyYAML>=6.0,<6.1